Pull Request #51: Add NullCharacterHandler
I am adding a NullCharacterHandler passed into the Row. The Reasoning
is, that I don't want to add methods like "setTextWithoutNullCharacter"
to the API. We can abstract this away nicely behind a handler, that the
library user can pass into the API. At the same time there is also a
"default" handler to replace 0x00 characters with an empty String.