Issue #52 Added a getLength(...) to IValueHandler
This is a breaking change, so at the same time I am also increasing the
pom.xml to 4.0. The getLength(...) is currently only used in the
Range<...> implementation, but it might be used for other
implementations as well. Some of the ValueHandlers throw a
NotImplementedException, because I am not planning to allow them for a
Range at the moment.