Issue #58 Add Annotation Processor
This implements a PostgresDataType Annotation, so the DataTypes can be
configured without passing additional data into the mapping, though it
is still possible to go low level without Annotations.

There is still a little bit of duplication, because the
Column-Annotation already contains the Column name. But for now I think
this helps to later extract the annotation into a common namespace
possibly reused by other modules.